+++ llama/configuration_llama.py
@@ -21,9 +21,7 @@
 
 from transformers.configuration_utils import PretrainedConfig
 from transformers.utils import logging
-
-
-logger = logging.get_logger(__name__)
+from loguru import logger
 
 LLAMA_PRETRAINED_CONFIG_ARCHIVE_MAP = {}