+++ llama/module/norm.py
@@ -16,5 +16,6 @@ class LlamaRMSNorm(nn.Module):
         hidden_states = hidden_states.to(torch.float32)
         variance = hidden_states.pow(2).mean(-1, keepdim=True)
         hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
+
         return self.weight * hidden_states.to(input_dtype)