+++ changelog.md
@@ -9,6 +9,25 @@ and this project adheres to [Semantic Versioning].
 
 - /
 
+
+
+## [0.0.3] - 2023-09-12
+
+### Added
+
+- utils.misc
+  - vis_model_stucture
+  - set_logger
+
+### Changed
+
+- llama/module
+  - attetnion: 注释`维度`和`前向过程`
+  - decorder layer: 增加层数
+  - mlp: 注释`维度`和`前向过程`
+  - model: 给 decoder layer 增加层数
+  - position embedding: 注释`维度`和`前向过程`
+
 ## [0.0.2] - 2023-09-11
 
 ### Changed
@@ -37,10 +56,12 @@ and this project adheres to [Semantic Versioning].
   - `llama-2-13b-chat-hf` 模型权重，convert_llama_weights_to_hf
 
 <!-- Links -->
+
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
+
 [unreleased]: https://github.com/Author/Repository/compare/v0.0.2...HEAD
 [0.0.2]: https://github.com/Author/Repository/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Author/Repository/releases/tag/v0.0.1